+++ tests/inputs-luau/excess-parentheses.lua
@@ -7,3 +7,12 @@ local foo = (if true then 0 else 1) + 1
 local firstPendingUpdate = ((lastPendingUpdate.next :: any) :: Update<State>)
 
 local x = #(value :: Array<number>)
+
+-- https://github.com/JohnnyMorganz/StyLua/issues/425
+self.mutationStore[mutationId] = (
+	{
+		lolz = foreva,
+		variables = variables,
+	} :: anyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy
+) :: MutationStoreValue
+